pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-ai-gateway"
version = "0.1.0"
description = "Production-grade AI gateway with asynchronous guardrails and streaming-first middleware"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "AI Team"},
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.25.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=0.41b0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-httpx>=0.41b0",
    "opentelemetry-exporter-prometheus>=0.41b0",
    "presidio-analyzer>=2.2.35",
    "presidio-anonymizer>=2.2.35",
    "spacy>=3.7.0",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "black>=23.12.0",
]

[tool.setuptools]
packages = ["sentinel"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=sentinel --cov-report=term-missing"
